// types/sandai_type.ts
enum SandaiType {
  WAGYU = "wagyu",
  CHINA_TOWN = "china-town",
  RED_LIGHT = "red-light",
  GASSHO = "gassho",
  CASTLE = "castle",
  RAMEN = "ramen",
  VIEW = "view",
  GARDEN = "garden",
  TEA = "tea",
  ONSEN = "onsen",
  NIGHT_VIEW = "night-view",
  SHRINE = "shrine",
  FIREWORK = "firework",
  TENMANGU = "tenmangu",
  FESTIVAL = "festival",
}

export default SandaiType;